Improve dataclass organization, remove older FROG_result
